requirements.txt
deepface==0.0.79
keras==2.14.0
matplotlib==3.5.3
numpy==1.19.3
opencv_python==4.1.0.25
pandas==1.3.5
scikit_learn==1.0.2
tensorflow_gpu==2.4.0
torch==2.1.0
tqdm==4.65.0
transformers==4.34.1
